refactor(gsd): route domain writes through writer layer
