+++ src/resources/extensions/gsd/db/writers/task-execution.ts
@@ -5,6 +5,103 @@ import type { DomainOperationContext } from "../domain-operation.js";
 import { getDb } from "../engine.js";
 import { requireActiveDomainOperationContext } from "./lifecycle-commands.js";
 
+export interface StagedTaskCompletionWriteInput {
+  task: {
+    milestoneId: string;
+    sliceId: string;
+    taskId: string;
+  };
+  oneLiner: string;
+  narrative: string;
+  verificationResult: string;
+  blockerDiscovered: boolean;
+  deviations: string;
+  knownIssues: string;
+  keyFiles: string[];
+  keyDecisions: string[];
+  fullSummaryMd: string;
+  verificationEvidence: Array<{
+    command: string;
+    exitCode: number;
+    verdict: string;
+    durationMs: number;
+  }>;
+}
+
+export function writeStagedTaskCompletion(
+  context: Readonly<DomainOperationContext>,
+  attempt: Readonly<{ milestoneId: string; sliceId: string; taskId: string }>,
+  completion: Readonly<StagedTaskCompletionWriteInput>,
+): void {
+  if (requireActiveDomainOperationContext(context) !== "attempt.settle") {
+    throw new Error("Staged Task completion requires an attempt.settle Domain Operation");
+  }
+  if (
+    completion.task.milestoneId !== attempt.milestoneId ||
+    completion.task.sliceId !== attempt.sliceId ||
+    completion.task.taskId !== attempt.taskId
+  ) {
+    throw new Error("Staged Task completion does not match the settlement Attempt");
+  }
+
+  const updated = getDb().prepare(`
+    UPDATE tasks
+    SET status = 'in_progress',
+        completed_at = NULL,
+        one_liner = :one_liner,
+        narrative = :narrative,
+        verification_result = :verification_result,
+        blocker_discovered = :blocker_discovered,
+        deviations = :deviations,
+        known_issues = :known_issues,
+        key_files = :key_files,
+        key_decisions = :key_decisions,
+        full_summary_md = :full_summary_md
+    WHERE milestone_id = :milestone_id
+      AND slice_id = :slice_id
+      AND id = :task_id
+      AND status NOT IN ('complete', 'done', 'closed')
+  `).run({
+    ":milestone_id": completion.task.milestoneId,
+    ":slice_id": completion.task.sliceId,
+    ":task_id": completion.task.taskId,
+    ":one_liner": completion.oneLiner,
+    ":narrative": completion.narrative,
+    ":verification_result": completion.verificationResult,
+    ":blocker_discovered": completion.blockerDiscovered ? 1 : 0,
+    ":deviations": completion.deviations,
+    ":known_issues": completion.knownIssues,
+    ":key_files": JSON.stringify(completion.keyFiles),
+    ":key_decisions": JSON.stringify(completion.keyDecisions),
+    ":full_summary_md": completion.fullSummaryMd,
+  });
+  if (Number((updated as { changes?: unknown }).changes ?? 0) !== 1) {
+    throw new Error("Staged Task completion target is missing or already complete");
+  }
+
+  const insertEvidence = getDb().prepare(`
+    INSERT OR IGNORE INTO verification_evidence (
+      task_id, slice_id, milestone_id, command, exit_code, verdict, duration_ms, created_at
+    )
+    SELECT :task_id, :slice_id, :milestone_id, :command, :exit_code, :verdict,
+           :duration_ms, operation.created_at
+    FROM workflow_operations operation
+    WHERE operation.operation_id = :operation_id
+  `);
+  for (const evidence of completion.verificationEvidence) {
+    insertEvidence.run({
+      ":task_id": completion.task.taskId,
+      ":slice_id": completion.task.sliceId,
+      ":milestone_id": completion.task.milestoneId,
+      ":command": evidence.command,
+      ":exit_code": evidence.exitCode,
+      ":verdict": evidence.verdict,
+      ":duration_ms": evidence.durationMs,
+      ":operation_id": context.operationId,
+    });
+  }
+}
+
 export interface TaskDispatchIdentity {
   dispatchId: number;
   workerId: string;
